requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
cssselect==1.1.0
cssutils==2.3.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap4==21.2
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.7.1
django-jazzmin==2.4.9
django-jquery==3.1.0
django-simple-history==3.0.0
emails==0.6
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
premailer==3.10.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.8
validate-email==1.3
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
